module_a_vision/requirements.txt
opencv-python>=4.9.0.80
numpy>=1.26.4
scikit-image>=0.22.0
scikit-learn>=1.4.0
Pillow>=10.2.0
imagehash>=4.3.1
matplotlib>=3.8.2
plotly>=5.18.0
scipy>=1.12.0
fastapi>=0.116.1
uvicorn[standard]>=0.27.1
python-multipart>=0.0.9
pydantic>=2.10.3
pydantic-settings>=2.1.0
httpx>=0.28.1
ipfshttpclient==0.8.0a2
PyYAML>=6.0.2
python-dotenv>=1.0.1
aiofiles>=23.2.1
tqdm>=4.66.1
pandas>=2.2.0
seaborn>=0.13.2
pytest>=7.4.4
pytest-asyncio>=0.23.4
pytest-cov>=4.1.0
flake8>=7.0.0
black>=24.1.1
